change Activities to activity, fragments to fragment
