+++ app/src/main/AndroidManifest.xml
@@ -15,19 +15,22 @@
         android:theme="@style/Theme.Truyenchu"
         tools:targetApi="31">
         <activity
-            android:name=".Activities.EditProfileActivity"
+            android:name=".activity.DonateActivity"
             android:exported="false" />
         <activity
-            android:name=".Activities.ChangePasswordActivity"
+            android:name=".activity.EditProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".Activities.ForgotPasswordActivity"
+            android:name=".activity.ChangePasswordActivity"
             android:exported="false" />
         <activity
-            android:name=".Activities.SigninActivity"
+            android:name=".activity.ForgotPasswordActivity"
             android:exported="false" />
         <activity
-            android:name=".Activities.SignupActivity"
+            android:name=".activity.SigninActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activity.SignupActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
