+++ app/src/main/java/com/example/truyenchu/activity/ChangePasswordActivity.java
@@ -1,4 +1,4 @@
-package com.example.truyenchu.Activities;
+package com.example.truyenchu.activity;
 
 import android.app.ProgressDialog;
 import android.content.Intent;
