+++ app/src/main/java/com/example/truyenchu/activity/EditProfileActivity.java
@@ -1,11 +1,9 @@
-package com.example.truyenchu.Activities; // Thay bằng package của bạn
+package com.example.truyenchu.activity; // Thay bằng package của bạn
 
 import android.app.ProgressDialog;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.util.Log;
-import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.Toast;
