+++ app/src/main/java/com/example/truyenchu/activity/ForgotPasswordActivity.java
@@ -1,11 +1,10 @@
-package com.example.truyenchu.Activities;
+package com.example.truyenchu.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Button;
 import android.widget.ImageView;
 
-import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.app.ProgressDialog;
