+++ app/src/main/java/com/example/truyenchu/fragment/CategoryFragment.java
@@ -1,4 +1,4 @@
-package com.example.truyenchu.fragments;
+package com.example.truyenchu.fragment;
 
 import android.os.Bundle;
 
@@ -12,10 +12,10 @@ import com.example.truyenchu.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link CommunityFragment#newInstance} factory method to
+ * Use the {@link CategoryFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class CommunityFragment extends Fragment {
+public class CategoryFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class CommunityFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public CommunityFragment() {
+    public CategoryFragment() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class CommunityFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment CommunityFragment.
+     * @return A new instance of fragment CategoryFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static CommunityFragment newInstance(String param1, String param2) {
-        CommunityFragment fragment = new CommunityFragment();
+    public static CategoryFragment newInstance(String param1, String param2) {
+        CategoryFragment fragment = new CategoryFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class CommunityFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_community, container, false);
+        return inflater.inflate(R.layout.fragment_category, container, false);
     }
 }
