+++ app/src/main/java/com/example/truyenchu/fragment/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.example.truyenchu.fragments;
+package com.example.truyenchu.fragment;
 
 import android.annotation.SuppressLint;
 import android.os.Bundle;
