+++ app/src/main/java/com/example/truyenchu/fragment/ProfileFragment.java
@@ -1,4 +1,4 @@
-package com.example.truyenchu.fragments;
+package com.example.truyenchu.fragment;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,6 +6,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -14,8 +15,9 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import com.bumptech.glide.Glide;
-import com.example.truyenchu.Activities.EditProfileActivity;
-import com.example.truyenchu.Activities.SigninActivity;
+import com.example.truyenchu.activity.DonateActivity;
+import com.example.truyenchu.activity.EditProfileActivity;
+import com.example.truyenchu.activity.SigninActivity;
 import com.example.truyenchu.R;
 import com.example.truyenchu.model.User;
 import com.google.firebase.auth.FirebaseAuth;
@@ -31,6 +33,7 @@ public class ProfileFragment extends Fragment {
     private CircleImageView avatarIv;
     private TextView nameTv, emailTv, editTv, logoutTv;
     private LinearLayout profileHeaderLayout;
+    private RelativeLayout donate;
 
     private FirebaseAuth mAuth;
     private FirebaseFirestore db;
@@ -51,6 +54,7 @@ public class ProfileFragment extends Fragment {
         editTv = view.findViewById(R.id.profile_edit_tv);
         logoutTv = view.findViewById(R.id.profile_logout_tv);
         profileHeaderLayout = view.findViewById(R.id.profile_header_layout);
+        donate = view.findViewById(R.id.donate);
 
         return view;
     }
@@ -67,6 +71,13 @@ public class ProfileFragment extends Fragment {
         logoutTv.setOnClickListener(null);
         profileHeaderLayout.setOnClickListener(null);
         editTv.setOnClickListener(null);
+
+
+        donate.setOnClickListener(v -> {
+            if (getContext() != null) {
+                startActivity(new Intent(getContext(), DonateActivity.class));
+            }
+        });
 
         FirebaseUser firebaseUser = mAuth.getCurrentUser();
 
@@ -106,7 +117,6 @@ public class ProfileFragment extends Fragment {
                     startActivity(new Intent(getContext(), EditProfileActivity.class));
                 }
             });
-
         } else {
             // --- TRƯỜNG HỢP: CHƯA ĐĂNG NHẬP (KHÁCH) ---
             nameTv.setText("Đăng nhập/Đăng ký");
