+++ app/src/main/res/layout/activity_change_password.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/screen_background"
-    tools:context=".Activities.ChangePasswordActivity">
+    tools:context=".activity.ChangePasswordActivity">
 
     <ImageView
         android:id="@+id/backToProfile"
