+++ app/src/main/res/layout/activity_forget_password.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/screen_background"
-    tools:context=".Activities.ForgotPasswordActivity">
+    tools:context=".activity.ForgotPasswordActivity">
 
     <ImageView
         android:id="@+id/backToSignin"
