+++ app/src/main/res/layout/activity_signin.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/screen_background"
     android:fillViewport="true"
-    tools:context=".Activities.SigninActivity">
+    tools:context=".activity.SigninActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
