+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,31 +3,31 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/navigation_home">
+    app:startDestination="@id/nav_home">
 
     <fragment
-        android:id="@+id/navigation_home"
-        android:name="com.example.truyenchu.fragments.HomeFragment"
+        android:id="@+id/nav_home"
+        android:name="com.example.truyenchu.fragment.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
     <fragment
-        android:id="@+id/navigation_category"
-        android:name="com.example.truyenchu.fragments.CategoryFragment"
+        android:id="@+id/nav_categories"
+        android:name="com.example.truyenchu.fragment.CategoryFragment"
         android:label="fragment_category"
         tools:layout="@layout/fragment_category" />
     <fragment
-        android:id="@+id/navigation_community"
-        android:name="com.example.truyenchu.fragments.CommunityFragment"
+        android:id="@+id/nav_community"
+        android:name="com.example.truyenchu.fragment.CommunityFragment"
         android:label="fragment_community"
         tools:layout="@layout/fragment_community" />
     <fragment
-        android:id="@+id/navigation_library"
-        android:name="com.example.truyenchu.fragments.LibraryFragment"
+        android:id="@+id/nav_library"
+        android:name="com.example.truyenchu.fragment.LibraryFragment"
         android:label="fragment_library"
         tools:layout="@layout/fragment_library" />
     <fragment
-        android:id="@+id/navigation_profile"
-        android:name="com.example.truyenchu.fragments.ProfileFragment"
+        android:id="@+id/nav_profile"
+        android:name="com.example.truyenchu.fragment.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
