+++ app/src/main/res/values/colors.xml
@@ -18,6 +18,7 @@
     <color name="text_hint">#9CA3AF</color>
     <color name="divider_color">#E5E7EB</color>
     <color name="transparent">#00000000</color>
+    <color name="donate">#2ac737</color>
 
 
 </resources>
